Refactored project.
Changed state -> editor.
Changed events -> tools
Refactored and commented tools.
Refactored and commented editor.
Changed mouse input handling for more consistency.
