+++ src/tools/prelude.rs
@@ -1,12 +1,12 @@
 // Our stuff
-pub use super::MouseMeta;
-pub use super::{center_cursor, mode_switched, update_mousepos, update_viewport};
+pub use super::{_center_cursor, update_viewport};
+pub use super::{EditorEvent, MouseEventType, MouseInfo, Tool};
 
 pub use crate::renderer::constants::*;
 pub use crate::renderer::points::calc::*;
-pub use crate::state;
-pub use crate::state::{PointData};
-pub use crate::util;
+pub use crate::editor;
+pub use crate::editor::{Editor, PointData, clicked_point_or_handle, SelectPointInfo, nearest_point_on_curve};
+pub use crate::util::*;
 pub use crate::{CONSOLE};
 
 // Skia/Winit stuff
